// frontend/src/types/models/ScenarioInitGeneral.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { TemperatureEnableEnum } from './TemperatureEnableEnum';
export type ScenarioInitGeneral = {
  temperature?: number;
  temperature_enable?: TemperatureEnableEnum;
};