// frontend/src/types/models/RhythmEnum.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
 
/**
 * * `sinus` - Sinus rhythm
 * * `afib` - Atrial fibrillation
 * * `vfib` - Ventricular fibrillation
 * * `vtach1` - Ventricular tachycardia 1
 * * `vtach2` - Ventricular tachycardia 2
 * * `ront` - R-on-T
 * * `asystole` - Asystole
 */
export enum RhythmEnum {
  SINUS = 'sinus',
  AFIB = 'afib',
  VFIB = 'vfib',
  VTACH1 = 'vtach1',
  VTACH2 = 'vtach2',
  RONT = 'ront',
  ASYSTOLE = 'asystole',
}