// frontend/src/types/models/VpcEnum.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */

/**
 * * `none` - None
 * * `1-1` - Vtach1 Singlet
 * * `1-2` - Vtach1 Couplet
 * * `1-3` - Vtach1 Triplet
 * * `2-1` - Vtach2 Singlet
 * * `2-2` - Vtach2 Couplet
 * * `2-3` - Vtach2 Triplet
 * * `3-1` - Vtach3 Singlet
 * * `3-2` - Vtach3 Couplet
 * * `3-3` - Vtach3 Triplet
 */
export enum VpcEnum {
  NONE = 'none',
  _1_1 = '1-1',
  _1_2 = '1-2',
  _1_3 = '1-3',
  _2_1 = '2-1',
  _2_2 = '2-2',
  _2_3 = '2-3',
  _3_1 = '3-1',
  _3_2 = '3-2',
  _3_3 = '3-3',
}